Compiled executable with embedded code
Added ./programs/buildlauncher.py. This module builds exelauncher.py and
compiles it to an executable. The executable includes the source code
for various modules (base/metaclass/network/etc) embedded as strings.
Combined with an Import hook and Cython, the resulting exe has the
source code for it's own logic embedded in itself as clear text.

The update and save/load mechanisms function, though for anything to
actually update the embedded strings will have to be made modifiable
and/or an optional source argument will need to be added to the update
method.