Authenticated Services updated
Mostly finished revamping to utilize srp and the new
Authenticated_Service/Client components.

Note that their is still a major security issue. Usage of the
RPC_Handler for the login process does not yet have authentication on
the pickled data and could be tampered with en route. Please do not use
it yet on machines connected to the internet!